fix: un-deprecated `.log()` calls to ease transition to Probot v10
