+++ test/deprecations.test.ts
@@ -3,7 +3,7 @@ import Stream from "stream";
 import { Webhooks } from "@octokit/webhooks";
 import pino from "pino";
 
-import { Application, Context, createProbot, Probot } from "../src";
+import { createProbot, Probot } from "../src";
 
 describe("Deprecations", () => {
   let output: any;
@@ -60,39 +60,6 @@ describe("Deprecations", () => {
     expect(output.length).toEqual(2);
     expect(output[0].msg).toContain(
       `[probot] "probot.logger" is deprecated. Use "probot.log" instead`
-    );
-  });
-
-  it("probot.log()", () => {
-    const probot = new Probot({ log: pino(streamLogsToOutput) });
-    probot.log("test");
-
-    expect(output.length).toEqual(2);
-    expect(output[0].msg).toContain(
-      '[probot] "app.log()" and "context.log()" are deprecated. Use "app.log.info()" and "context.log.info()" instead'
-    );
-  });
-
-  it("app.log()", () => {
-    const app = new Application({
-      log: pino(streamLogsToOutput),
-      secret: "secret",
-    });
-    app.log("test");
-
-    expect(output.length).toEqual(2);
-    expect(output[0].msg).toContain(
-      '[probot] "app.log()" and "context.log()" are deprecated. Use "app.log.info()" and "context.log.info()" instead'
-    );
-  });
-
-  it("context.log()", () => {
-    const context = new Context({} as any, {} as any, pino(streamLogsToOutput));
-    context.log("test");
-
-    expect(output.length).toEqual(2);
-    expect(output[0].msg).toContain(
-      '[probot] "app.log()" and "context.log()" are deprecated. Use "app.log.info()" and "context.log.info()" instead'
     );
   });
 });
